Fixed memory leak, removed unneeded hook
